+++ shop-backend/src/main/kotlin/ru/morsianin_shop/model/Order.kt
@@ -12,7 +12,7 @@ data class Order(
     @Serializable
     data class Item(
         val productId: Long,
-        val quantity: Long
+        val quantity: Int
     )
 
     @Serializable
